backend/requirements.txt
# Zenithai Backend - Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Config & env
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0

# Redis (optional, for cache/queue)
redis==5.0.1

# Market data & trading (crypto)
ccxt==4.2.25
pandas==2.2.0
numpy==1.26.4

# ML
scikit-learn==1.4.0
joblib==1.4.0
